Add media player properties
